Move AssemblyDuplicateReferences back to ExpressionCompilerTests
(cherry picked from commit f27c7d0cc248f9d808ea16b9a8c8223c5d06cce9)
