+++ src/ExpressionEvaluator/CSharp/Test/ExpressionCompiler/CSharpExpressionCompilerTest.csproj
@@ -96,7 +96,6 @@
     <Compile Include="DebuggerDisplayAttributeTests.cs" />
     <Compile Include="DeclarationTests.cs" />
     <Compile Include="ExpressionCompilerTestBase.cs" />
-    <Compile Include="MetadataUtilitiesTests.cs" />
     <Compile Include="MissingAssemblyTests.cs" />
     <Compile Include="InstructionDecoderTests.cs" />
     <Compile Include="HoistedStateMachineLocalTests.cs" />
